+++ libnavigation-base/src/test/java/com/mapbox/navigation/base/route/RouteExclusionsTest.kt
@@ -35,6 +35,7 @@ class RouteExclusionsTest {
                     add(
                         mockk {
                             every { routeId } returns "$it"
+                            every { routerOrigin } returns com.mapbox.navigator.RouterOrigin.ONBOARD
                         }
                     )
                 }